Build script added; internal ode removed
Due to some unknown phenomenon, the internal ode refuses to
compile without some additional effort. This is despite the whole
folder being bitwise identical to a freshly unpacked one from
ode-0.12.tar.gz. In the interest of time, I'm removing an internal
ODE and replacing it with a build script which automatically
downloads the file.

	new file:   build.sh
	deleted recursively:    ode-0.12/