Screen capture compositing investigated and switched to mode compartible with Linux OpenGL rendering
The comment on lines 145-149 of pyrosim/simulator/simulator.cpp:

FIXME: It appears to be a difference in pixel formats between MacOS and Linux
(and I vaguely remember that having something to do with compositing).
Figure out a way to autodetect the format and choose the right one for reading.
Useful reading on glGet:
https://www.khronos.org/opengl/wiki/GLAPI/glGet#Framebuffers

	modified:   demos/ludobots/Demo_22_Screen_Capture.py
	deleted:    frame/2017-11-15-20-24-16_movie.mp4
	modified:   pyrosim/simulator/simulator.cpp